@@ -2,7 +2,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'io.fabric'
 
 android {
-	compileSdkVersion 27
+	compileSdkVersion 28
 	buildToolsVersion '28.0.3'
 
 	compileOptions {
@@ -13,7 +13,7 @@ android {
 	defaultConfig {
 		applicationId 'org.openobservatory.ooniprobe'
 		minSdkVersion 18
-		targetSdkVersion 27
+		targetSdkVersion 28
 		versionName '2.0.0-alpha.6'
 		versionCode 32
 	}
@@ -28,10 +28,10 @@ android {
 }
 
 dependencies {
-	implementation 'com.android.support:appcompat-v7:27.1.1'
-	implementation 'com.android.support:support-v13:27.1.1'
-	implementation 'com.android.support:cardview-v7:27.1.1'
-	implementation 'com.android.support:preference-v7:27.1.1'
+	implementation 'androidx.appcompat:appcompat:1.0.0'
+	implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+	implementation 'androidx.cardview:cardview:1.0.0'
+	implementation 'androidx.preference:preference:1.0.0'
 
 	implementation 'com.google.firebase:firebase-messaging:17.3.2'
 	implementation 'com.google.firebase:firebase-core:16.0.3'
@@ -43,12 +43,12 @@ dependencies {
 	implementation 'com.squareup.retrofit2:retrofit:2.4.0'
 	implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
 
-	annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
-	implementation 'com.jakewharton:butterknife:8.8.1'
+	annotationProcessor 'com.jakewharton:butterknife-compiler:9.0.0-SNAPSHOT'
+	implementation 'com.jakewharton:butterknife:9.0.0-SNAPSHOT'
 
 	implementation('com.crashlytics.sdk.android:crashlytics:2.9.3@aar') { transitive = true }
 	implementation 'org.openobservatory.measurement_kit:android-libs:0.9.0-alpha.9-1-android.2'
-	implementation 'com.github.xanscale.LocalhostToolkit:app:b1cbb84c43'
+	implementation 'com.github.xanscale.LocalhostToolkit:app:0cc7be9da7'
 	implementation 'pub.devrel:easypermissions:1.3.0'
 	implementation 'com.airbnb.android:lottie:2.5.6'
 	implementation 'com.github.apl-devs:appintro:v4.2.3'
