@@ -3,7 +3,7 @@ package org.openobservatory.ooniprobe.activity;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.view.Window;
 
 import org.openobservatory.ooniprobe.R;
