@@ -2,7 +2,7 @@ package org.openobservatory.ooniprobe.activity;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import org.openobservatory.ooniprobe.BuildConfig;
