@@ -3,7 +3,7 @@ package org.openobservatory.ooniprobe.activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.animation.Animation;
 import android.widget.ProgressBar;
 import android.widget.TextView;
