@@ -3,8 +3,8 @@ package org.openobservatory.ooniprobe.fragment.consent;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v7.widget.AppCompatButton;
+import androidx.fragment.app.DialogFragment;
+import androidx.appcompat.widget.AppCompatButton;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
