@@ -2,7 +2,7 @@ package org.openobservatory.ooniprobe.fragment.resultHeader;
 
 import android.app.Fragment;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
