@@ -1,7 +1,7 @@
 package org.openobservatory.ooniprobe.test.test;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.gson.Gson;
 
