@@ -1,6 +1,6 @@
 package org.openobservatory.ooniprobe.utils;
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
 
